Pull method out into its own util
Up next is to test
