+++ src/services/asset-discovery-service.ts
@@ -5,47 +5,12 @@ import { AssetDiscoveryConfiguration } from '../configuration/asset-discovery-co
 import { DEFAULT_CONFIGURATION } from '../configuration/configuration'
 import { SnapshotOptions } from '../percy-agent-client/snapshot-options'
 import { logError, profile } from '../utils/logger'
+import { cacheResponse, getResponseCache } from '../utils/response-cache'
 import waitForNetworkIdle from '../utils/wait-for-network-idle'
 import PercyClientService from './percy-client-service'
 import ResponseService from './response-service'
 
 export const MAX_SNAPSHOT_WIDTHS: number = 10
-const requestCache = {} as any
-
-/**
- * Keep an in-memory cache of asset responses.
- *
- * When enabled, asset responses will be kept in memory. When the asset is
- * re-requested, it will be responsed with what the cached response. This makes
- * it so servers aren't being hounded for the same asset over and over again.
- */
-async function cacheResponse(response: puppeteer.Response, logger: any) {
-  const responseUrl = response.url()
-  const statusCode = response.status()
-
-  if (!!requestCache[responseUrl]) {
-    logger.debug(`Asset already in cache ${responseUrl}`)
-    return
-  }
-
-  if (![200, 201].includes(statusCode)) {
-    return
-  }
-
-  try {
-    const buffer = await response.buffer()
-
-    requestCache[responseUrl] = {
-      status: response.status(),
-      headers: response.headers(),
-      body: buffer,
-    }
-
-    logger.debug(`Added ${responseUrl} to asset discovery cache`)
-  } catch (error) {
-    logger.debug(`Could not cache response ${responseUrl}: ${error}`)
-  }
-}
 
 export class AssetDiscoveryService extends PercyClientService {
   responseService: ResponseService
@@ -242,9 +207,9 @@ export class AssetDiscoveryService extends PercyClientService {
           return
         }
 
-        if (this.configuration['cache-responses'] === true && requestCache[requestUrl]) {
+        if (this.configuration['cache-responses'] === true && getResponseCache()[requestUrl]) {
           logger.debug(`Asset cache hit for ${requestUrl}`)
-          await request.respond(requestCache[requestUrl])
+          await request.respond(getResponseCache()[requestUrl])
 
           return
         }
